Use x/sys/unix epoll (#737)
* Make gobgp compile in an image

This patch adds Makefile logic, similar to the kube-router target,
allowing gobgp to be built in a container.

* Use unix.Epoll* functions

To be able to compile and run on the Linux@arm64 architecture
one has to use the `golang.org/x/sys/unix` package instead
the `syscall` package. This is because of these Go upstream
bugs that won't be fixed in the standard library:

- syscall: arm64: epoll_wait syscall not implemented
  https://github.com/golang/go/issues/25813

(cherry picked from commit 6470795db3c29dc3d54ff4a347476c4a5d2ce22a)